add, delete expert validation and result return

+++ src/models/experts.py
@@ -1,4 +1,5 @@
 from helpers.database import get_mysql_connection as get_db
+from helpers.result import OperationResult as Result
 
 class Expert:
     def __init__(self, name, mail, id=None):
@@ -20,18 +21,25 @@ def get_experts() -> list:
     db.close()
     return experts
 
-def create_expert(expert: Expert) -> None:
+def create_expert(expert: Expert) -> Result:
     db = get_db()
     cursor = db.cursor()
     cursor.execute('INSERT INTO Experts (name, address) VALUES (%s, %s)', (expert.name, expert.mail))
     db.commit()
     cursor.close()
     db.close()
+    return Result(True, "Expert created successfully")
     
-def delete_expert(expert_id: int) -> None:
+def delete_expert(expert_id: int) -> Result:
     db = get_db()
     cursor = db.cursor()
+    
+    cursor.execute('SELECT * FROM Model_Experts WHERE expert_id = %s', (expert_id,))
+    if cursor.fetchone() is not None:
+        return Result(False, 'Expert is being used in a model!')
+    
     cursor.execute('DELETE FROM Experts WHERE expert_id = %s', (expert_id,))
     db.commit()
     cursor.close()
     db.close()
+    return Result(True, "Expert deleted successfully")
